Crate nnom

A minimal presentation of a nom-like parser combinator framework.

Heavily inspired by synom, but using only the parts I need.

Usage

A parser fragment takes the form fn(input: In) -> Result<In, Out, Error>. On a successful parse, a ParseResult is returned, which contains the unused slice of the input and the parsed value. On a failed parse, a hopefully descriptive error is provided.

Here In stands in for some slice that you work on. Out is commonly the matched slice, or some transformed form, such as a token or AST. This crate does not prescribe error handling, but suggests that using error-chain is beneficial for passing errors up the parser chain.

Modules

combinators
Simple combinator fn of other parser fn

prelude
The commonly used types of this crate

result
Parsing result types

slice
Types for working with slices while retaining position information